+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Deprecated `next_id` function in favor of `new_id`
 - Added `identity` and `has_identity` functions
 - Added `singleton` and `events` functions
+- Added `select_all` function
 
 ## 1.1.0 (September 12, 2017)