+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 2.0.0.dev1 (Future release)
+## 2.0.0 (March 19, 2018)
 
 - Added support for recursive references
 - Added `abort` function